requirements.txt
absl-py==2.1.0
astor==0.8.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
gast==0.2.2
google-pasta==0.2.0
grpcio==1.62.2
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.4
MarkupSafe==2.1.5
neuralgym @ git+https://github.com/JiahuiYu/neuralgym@88292adb524186693a32404c0cfdc790426ea441
numpy==1.18.5
opencv-python==4.9.0.80
opt-einsum==3.3.0
Pillow==9.5.0
protobuf==3.20.0
PyYAML==6.0.1
six==1.16.0
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==2.3.0
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
